requirements.txt
streamlit
qdrant-client
google-generativeai
python-dotenv
uuid
pandas 